% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{trint_subs_labels}
\alias{trint_subs_labels}
\title{Shared functions for NanoSeqTools}
\format{
An object of class \code{character} of length 96.
}
\usage{
trint_subs_labels
}
\description{
Shared functions for NanoSeqTools
}
\keyword{datasets}
