% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nanoseq_data.R
\name{load_nanoseq_data}
\alias{load_nanoseq_data}
\title{Load all data required for NanoSeq analysis}
\usage{
load_nanoseq_data(
  dirs,
  suffix_to_remove = "NA",
  BSgenomepackagename,
  BSgenomecontigs
)
}
\arguments{
\item{dirs}{List of all the directories containing NanoSeq results to load.}

\item{suffix_to_remove}{A string to remove from the NanoSeq results directory names, to retain only the Sample IDs without any suffixes. If there is no suffix to remove, give the input as "NA" (default).}

\item{BSgenomepackagename}{A string of a BSgenome package corresponding to the reference genome used in the Nanoseq analysis (e.g., "BSgenome.Hsapiens.UCSC.hg38" or "BSgenome.Mmusculus.UCSC.mm10"). This is used to calculate the genome trinucleotide background and corrected substitution counts and burdens.}

\item{BSgenomecontigs}{A vector of numeric indices of the contigs of the BSgenome package from which to calculate the genome trinucleotide background (e.g., 1:24 for BSgenome.Hsapiens.UCSC.hg38, or 1:21 for BSgenome.Mmusculus.UCSC.mm10).}
}
\value{
A list of data files with all the relevant results generated by NanoSeq.
\itemize{
\item sample_id: A vector of all sample IDs that were loaded
\item vcf_snp.fix: List (one object per sample) containing the fixed information (fix) from the SNP vcf (only FILTER = PASS mutations)
\item vcf_indel.fix: List (one object per sample) containing the fixed information (fix) from the indel vcf (only FILTER = PASS mutations)
\item vcf_indel.gt: List (one object per sample) containing the genotype information (gt) from the indel vcf (only FILTER = PASS mutations)
\item trinuc_bg_ratio: Data frame of the sample trinucleotide background (i.e. number of interrogated bases for each trinucleotide context), the genome trinucleotide background (i.e. number of each trinucleotide context), and the normalized ratio of these. Columns: sample, tri (trinucleotide context), sample_tri_bg, genome_tri_bg, ratio2genome.
\item trinuc_bg.sigfit: Data frame in sigfit format of the trinucleotide background (i.e. number of interrogated bases for each trinucleotide context), with one row per sample and one column per trinucleotide context.
\item observed_corrected_trinuc_counts: Data frame of observed and corrected mutation counts (for all mutations and for unique mutations). Columns: sample, tri (trinucleotide context), trint_subst_observed, trint_subst_unique_observed, ratio2genome, trint_subst_corrected, trint_subst_unique_corrected.
\item observed_trinuc_counts.sigfit: Data frame in sigfit format of unique observed mutation counts, with one row per sample and one column per trinucleotide substitution context.
\item mutation_burden: The total number of observed and corrected mutations, total number of observed and corrected interrogated bases (note: observed and corrected are the same), observed and corrected mutation burdens, observed and corrected lower and upper confidence intervals of mutation counts, and observed and corrected lower and upper confidence intervals of mutation burdens, with one row per sample. All these statistics include all mutations, not just unique mutations.
\item purine_trinuc_mismatches: Data frame of the number of single-strand consensus purine mismatches. Columns: sample, tri (trinucleotide context), value.
\item pyrimidine_trinuc_mismatches: Data frame of the number of single-strand consensus pyrimidine mismatches. Columns: sample, tri (trinucleotide context), value.
\item estimated_error_rates: Data frame of the probability of having independent errors affecting both strands and resulting in a false-positive double-strand mutation and the number of estimated false positive double-strand mutations, based on the independent error rates in the purine channels. Columns: sample, total_error_rate, total_errors.
}
}
\description{
Package requirements (load before use): tidyverse, vcfR, BSgenome package corresponding to the reference genome used in the NanoSeq analysis
}
\details{
This function loads all the files for analysis of NanoSeq data. For more information regarding NanoSeq output files, refer to the "output" section in the \href{https://github.com/cancerit/NanoSeq}{NanoSeq GitHub}.
}
