% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nanoseq_data.R
\name{load_nanoseq_data}
\alias{load_nanoseq_data}
\title{Load NanoSeq data for genome-wide analysis}
\usage{
load_nanoseq_data(dirs, sample_names, BSgenomepackagename, BSgenomecontigs)
}
\arguments{
\item{dirs}{A character vector of the directories containing NanoSeq results to load (one directory per sample).}

\item{sample_names}{A character vector of the sample names to assign to the results, in the same order as the directories in 'dirs'.}

\item{BSgenomepackagename}{A string of a BSgenome package corresponding to the reference genome used in the Nanoseq analysis (e.g., "BSgenome.Hsapiens.UCSC.hg38" or "BSgenome.Mmusculus.UCSC.mm10"). This is used to calculate the genome trinucleotide background and corrected substitution counts and burdens.}

\item{BSgenomecontigs}{A vector of numeric indices of the contigs of the BSgenome package from which to calculate the genome trinucleotide background (e.g., 1:24 for BSgenome.Hsapiens.UCSC.hg38, or 1:21 for BSgenome.Mmusculus.UCSC.mm10).}
}
\value{
A list of data files with all the relevant results generated by NanoSeq.
\itemize{
\item sample_names: A vector of all sample IDs that were loaded
\item dirs: A character vector of the directories containing the NanoSeq results that were loaded
\item vcf_snp.fix: List (one object per sample) containing the fixed information (fix) from the SNP vcf (only FILTER = PASS mutations)
\item vcf_indel.fix: List (one object per sample) containing the fixed information (fix) from the indel vcf (only FILTER = PASS mutations)
\item vcf_indel.gt: List (one object per sample) containing the genotype information (gt) from the indel vcf (only FILTER = PASS mutations)
\item indel.spectra.sigfit: Data frame in sigfit format containing the indel spectra (created with indelwald indel.spectrum function), with one row per sample and one column per indel context
\item trinuc_bg_counts_ratio: Data frame of the sample trinucleotide background counts (i.e. number of interrogated bases for each trinucleotide context), the genome trinucleotide background counts (i.e. number of each trinucleotide context), and the normalized ratio of these. Columns: sample, tri (trinucleotide context), sample_tri_bg, genome_tri_bg, ratio2genome.
\item trinuc_bg_counts.sigfit: Data frame in sigfit format of the sample trinucleotide background counts, with one row per sample and one column per trinucleotide context.
\item trinuc_bg_ratio.sigfit: Data frame in sigfit format of the ratio of the sample trinucleotide background counts (normalized to a sum of 1) to the genome trinucleotide background counts (normalized to a sum of 1), with one row per sample and one column per trinucleotide context.
\item genome_trinuc_counts.sigfit: Vector of the genome trinucleotide background counts, in the same order as columns in sigfit format columns.
\item observed_corrected_trinuc_counts: Data frame of observed and corrected mutation counts (for all mutations and for unique mutations). Columns: sample, tri (trinucleotide context), trint_subst_observed, trint_subst_unique_observed, ratio2genome, trint_subst_corrected, trint_subst_unique_corrected.
\item observed_trinuc_counts.sigfit: Data frame in sigfit format of unique observed mutation counts, with one row per sample and one column per trinucleotide substitution context.
\item mutation_burden, with one row per sample. All substitution mutation statistics include all mutations, not just unique mutations.
}
\itemize{
\item The number of observed and corrected substitution mutations (muts_observed and muts_corrected)
\item Number of observed indels (indels_observed)
\item Total number of observed and corrected interrogated bases (total_observed and total_corrected; note: observed and corrected are the same)
\item Observed and corrected substitution mutation burdens (burden_observed and burden_corrected)
\item Observed indel mutation burden (burden_indels_observed)
\item Observed and corrected lower and upper confidence intervals of substitution mutation counts and observed lower and upper confidence intervals of indel counts (muts_lci_observed, muts_lci_corrected, indels_lci_observed, muts_uci_observed, muts_uci_corrected, indels_uci_observed)
\item Observed and corrected lower and upper confidence intervals of substitution mutation burdens and observed lower and upper confidence intervals of indel mutation burdens (burden_lci_observed, burden_lci_corrected, burden_indels_lci_observed, burden_uci_observed, burden_uci_corrected, burden_indels_uci_observed)
}
\itemize{
\item purine_trinuc_mismatches: Data frame of the number of single-strand consensus purine mismatches. Columns: sample, tri (trinucleotide context), value.
\item pyrimidine_trinuc_mismatches: Data frame of the number of single-strand consensus pyrimidine mismatches. Columns: sample, tri (trinucleotide context), value.
\item estimated_error_rates: Data frame of the probability of having independent errors affecting both strands and resulting in a false-positive double-strand mutation and the number of estimated false positive double-strand mutations, based on the independent error rates in the purine channels. Columns: sample, total_error_rate, total_errors.
}
}
\description{
Package requirements: tidyverse, vcfR, BSgenome package corresponding to the reference genome used in the NanoSeq analysis
}
\details{
This function loads NanoSeq data required for genome-wide analysis. For more information regarding NanoSeq output files, refer to the "output" section in the \href{https://github.com/cancerit/NanoSeq}{NanoSeq GitHub}.
}
