% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nanoseq_regions.R
\name{load_nanoseq_regions}
\alias{load_nanoseq_regions}
\title{Load NanoSeq data for region-specific analysis}
\usage{
load_nanoseq_regions(nanoseq_data, regions.list, tabix_bin)
}
\arguments{
\item{nanoseq_data}{Dataset resulting from load_nanoseq_data function}

\item{regions.list}{GRangesList object, comprised of GRanges that each contains a 'region set' to jointly analyze. The regions within each 'region set' can have overlaps (the functions handle this). Regions that were excluded when running load_nanoseq_data ('exclude_regions') are excluded from all region sets. The strand of each region in the region set specifies which mutations to include: '+ and '-' strand include mutations where central pyrimidine is on the '+' and '-' strands of the reference genome, respectively, and '*' includes all mutations. When there are overlapping regions with opposite strands within the same 'region set', the mutations in those overlapping regions are counted only once, because each mutation is a central pyrimidine on only one strand. Best practice is to name the elements of regions.list, since these names are carried forward to the output.}

\item{tabix_bin}{Full path of tabix binary}
}
\value{
Returns NanoSeq results for each 'region set'. Samples without coverage of any of the regions are skipped and absent from the output.
\itemize{
\item sample_names: A vector of all sample IDs that were loaded
\item dir: A vector of the directories containing the NanoSeq results that were loaded
\item regions.list: Copy of input regions.list
\item excluded_samples: Names of samples excluded from the results because they do not have NanoSeq read coverage in any regions
\item trinuc_bg_counts_ratio: Data frame of the sample trinucleotide background counts (i.e. for each sample and region set, the number of interrogated bases for each trinucleotide context), the genome trinucleotide background counts (i.e. for each sample/region set, the number of each trinucleotide context in the genomic region set), and the normalized ratio of these for each sample/region set combination. Columns: sample, region, tri (trinucleotide context), sample_tri_bg, genome_tri_bg, ratio2genome.
\item trinuc_bg_counts.sigfit: List with one object per region set, each comprised of a data frame in sigfit format of the sample trinucleotide background counts, with one row per sample and one column per trinucleotide context.
\item trinuc_bg_ratio.sigfit: List with one object per region set, each comprised of a data frame in sigfit format of the ratio of the sample trinucleotide background counts (normalized to a sum of 1) to the genome trinucleotide background counts (normalized to a sum of 1), with one row per sample and one column per trinucleotide context.
\item genome_trinuc_counts.sigfit: Vector of the genome trinucleotide background counts, in the same order as columns in sigfit format columns.
\item observed_corrected_trinuc_counts: Data frame of observed and corrected mutation counts (for all mutations and for unique mutations) for each sample/region set combination. Columns: sample, region, tri (trinucleotide context), trint_subst_observed, trint_subst_unique_observed, ratio2genome, trint_subst_corrected, trint_subst_unique_corrected.
\item observed_trinuc_counts.sigfit: List with one object per region set, each comprised of a data frame in sigfit format of unique observed mutation counts, with one row per sample and one column per trinucleotide substitution context.
\item mutation_burden: Data frame of the total number of observed and corrected mutations, total number of observed and corrected interrogated bases (note: observed and corrected are the same), observed and corrected mutation burdens, observed and corrected lower and upper confidence intervals of mutation counts, and observed and corrected lower and upper confidence intervals of mutation burdens, for each sample/region combination. All these statistics include all mutations, not just unique mutations.
}
}
\description{
Package requirements: tidyverse, rtracklayer, GenomicRanges, tabix binary
}
\details{
This function loads NanoSeq data required for region-specific analysis.
}
